Fix eglot integration with flycheck
As stated in the longer header of the file, now flymake and flycheck
live together in a buffer if eglot and :checkers syntax are active,
flycheck has no way to handle properly the updates eglot sends (at
least I, @gagbo, didn't find any), so now flycheck queries flymake for
the correct diagnostics to use, and an advice is used to trigger
flycheck on updates.

The implementation changed, triggered by an eglot change that fixed the
previous buggy behaviour (no updates were sent at all).